+++ NEWS.md
@@ -9,10 +9,14 @@ fastmap 1.1.0.9000
 
 * Fixed #27: If a `fastmap` object has no holes in the lists storing keys and values, and then it is serialized and then unserialized, the new `fastmap` would contain zero items. (#28)
 
+* Faster implementations of `fastmap` `$keys()` and `$as_list()` methods.
+
+
 fastmap 1.1.0
 =============
 
 * Added `faststack()` and `fastqueue()`. (#15)
+
 
 fastmap 1.0.1
 =============